Create NimbusOnboardingConfigUtility
